Declare W and H as long long int for 64b min. size
A c++11 type that guarantees minimum 64b size

+++ rtengine/labimage.h
@@ -29,7 +29,7 @@ private:
     void allocLab(size_t w, size_t h);
 
 public:
-    int W, H;
+    long long int W, H;
     float * data;
     float** L;
     float** a;
